# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "成果云系统API文档",
    "description": "成果云系统后端接口文档（Flask + MySQL）",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "登录后返回的 Bearer Token，格式：Bearer <access_token>"
    }
  },
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录",
        "description": "登录成功返回 access_token 与用户信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {"type": "string", "example": "admin"},
                "password": {"type": "string", "example": "123456"}
              }
            }
          }
        ],
        "responses": {
          "200": {"description": "登录成功，返回 token 和用户信息"},
          "401": {"description": "密码错误"},
          "404": {"description": "用户名不存在"}
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["认证"],
        "summary": "获取当前登录用户信息",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "返回当前用户信息（含角色与权限码）"},
          "401": {"description": "未登录或登录已过期"}
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "tags": ["认证"],
        "summary": "获取个人资料",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回个人资料"}}
      },
      "put": {
        "tags": ["认证"],
        "summary": "更新个人资料",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "body", "in": "body", "schema": {"type": "object"}}
        ],
        "responses": {"200": {"description": "更新成功"}}
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": ["认证"],
        "summary": "修改密码",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "old_password": {"type": "string"},
                "new_password": {"type": "string"}
              }
            }
          }
        ],
        "responses": {"200": {"description": "修改成功"}}
      }
    },
...
    "/api/register/phone": {
      "post": {
        "tags": ["注册"],
        "summary": "手机号验证码注册",
        "responses": {"200": {"description": "注册成功"}, "400": {"description": "参数错误"}}
      }
    },
...
    "/api/teachers/": {
      "get": {
        "tags": ["教师"],
        "summary": "教师列表（分页）",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回教师列表"}}
      },
      "post": {
        "tags": ["教师"],
        "summary": "添加教师",
        "security": [{"BearerAuth": []}],
        "responses": {"201": {"description": "添加成功"}}
      }
    },
...
  }
}